refactored wpf shadow to support any element implementing BaseViewManager
