+++ ReactWindows/ReactNative.Shared/Views/View/ReactViewManager.cs
@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using Newtonsoft.Json.Linq;
-using ReactNative.Reflection;
+﻿using ReactNative.Reflection;
 using ReactNative.UIManager;
 using ReactNative.UIManager.Annotations;
 #if WINDOWS_UWP
-using Windows.UI.Composition;
 using Windows.UI.Xaml.Controls;
 #else
 using System.Windows.Controls;
-using System.Windows.Media.Effects;
 #endif
 
 namespace ReactNative.Views.View
@@ -66,59 +61,6 @@ namespace ReactNative.Views.View
         protected override Canvas CreateInnerElement(ThemedReactContext reactContext)
         {
             return new Canvas();
-        }
-
-        [ReactProp(ViewProps.ShadowColor, CustomType = "Color")]
-        public void SetShadowColor(Border view, uint? color)
-        {
-#if WINDOWS_UWP
-            // TODO
-#else
-            DropShadowEffect effect = (DropShadowEffect) view.Effect ?? new DropShadowEffect();
-            effect.Color = ColorHelpers.Parse(color.Value);
-            view.Effect = effect;
-#endif
-        }
-
-        [ReactProp(ViewProps.ShadowOffset)]
-        public void SetShadowOffset(Border view, JObject offset)
-        {
-#if WINDOWS_UWP
-
-#else
-            DropShadowEffect effect = (DropShadowEffect) view.Effect ?? new DropShadowEffect();
-            var deltaX = offset.Value<double>("width");
-            var deltaY = offset.Value<double>("height");
-            var angle = Math.Atan2(deltaY, deltaX) * (180 / Math.PI);
-            var distance = Math.Sqrt(Math.Pow(deltaX, 2) + Math.Pow(deltaY, 2));
-            effect.Direction = angle;
-            effect.ShadowDepth = distance;
-            view.Effect = effect;
-#endif
-        }
-
-        [ReactProp(ViewProps.ShadowOpacity)]
-        public void SetShadowOpacity(Border view, double opacity)
-        {
-#if WINDOWS_UWP
-
-#else
-            DropShadowEffect effect = (DropShadowEffect)view.Effect ?? new DropShadowEffect();
-            effect.Opacity = opacity;
-            view.Effect = effect;
-#endif
-        }
-
-        [ReactProp(ViewProps.ShadowRadius)]
-        public void SetShadowRadius(Border view, double radius)
-        {
-#if WINDOWS_UWP
-
-#else
-            DropShadowEffect effect = (DropShadowEffect)view.Effect ?? new DropShadowEffect();
-            effect.BlurRadius = radius;
-            view.Effect = effect;
-#endif
         }
     }
 }
